requirements.txt
fastapi
pymongo
uvicorn
python-dotenv
pydantic
setuptools
pika
ffmpeg-python